dev: add component UniversalBitcoinAmountInput
