+++ src/components/Send/SendForm.tsx
@@ -8,7 +8,8 @@ import Sprite from '../Sprite'
 import { SourceJarSelector } from './SourceJarSelector'
 import { CoinjoinPreconditionViolationAlert } from '../CoinjoinPreconditionViolationAlert'
 import { DestinationInputField, DestinationValue } from './DestinationInputField'
-import { AmountInputField, AmountValue } from './AmountInputField'
+import { AmountInputField } from './AmountInputField'
+import { AmountValue } from '../UniversalBitcoinAmountInput'
 import CollaboratorsSelector from './CollaboratorsSelector'
 import { SweepBreakdown } from './SweepBreakdown'
 import FeeBreakdown from './FeeBreakdown'
